+++ lib/notifier.php
@@ -137,6 +137,17 @@ class Notifier implements INotifier {
 
 				$notification->setLink($editorLink);
 				break;
+			case "documentUnsaved":
+				$fileId = $parameters["fileId"];
+				$fileName = $parameters["fileName"];
+
+				$this->logger->info("Notify prepare: unsaved document $fileId");
+
+				$editorLink = $this->urlGenerator->linkToRouteAbsolute($this->appName . ".editor.index", ["fileId" => $fileId]);
+				$notification->setParsedSubject($trans->t("%1\$s could not be saved. Please open the file again.", [$fileName]))
+					->setIcon($this->urlGenerator->getAbsoluteURL($this->urlGenerator->imagePath($this->appName, 'app-dark.svg')));
+				$notification->setLink($editorLink);
+				break;
 			default:
 				$this->logger->info("Unsupported notification object: " . $notification->getObjectType(), ["app" => $this->appName]);
 		}