+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## Added
 - plugin description and useful links in admin settings
+- document unsaved notifications
 
 ## 9.10.1
 ## Added